backend/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
python-multipart
redis